adjusted file structure and modify profile layout
